pyproject.toml
[tool.poetry]
name = "taio-bill"
version = "0.1.0"
description = ""
authors = ["VIET-GITL <viet.info.43@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.22.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
pymupdf = "^1.24.7"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
